:recycle: [Components] Update renamed file imports
