+++ src/App.tsx
@@ -7,8 +7,8 @@ import {Footer} from '@src/components/sections/Footer/Footer.tsx';
 import {Header} from '@src/components/sections/Header/Header.tsx';
 import {Main} from '@src/components/sections/Main/Main.tsx';
 
-import {AccordionTest} from '@src/components/ui/Accordion/Accordion.test.tsx';
-import {ButtonTest} from '@src/components/ui/Button/Button.test.tsx';
+import {AccordionTest} from '@src/components/ui/Accordion/AccordionTest.tsx';
+import {ButtonTest} from '@src/components/ui/Button/ButtonTest.tsx';
 
 import styles from './App.module.css';
